requirements.txt
pandas==2.2.2
numpy==1.26.2
matplotlib==3.8.1
seaborn==0.13.2
scikit-learn==1.3.2
sqlalchemy==2.1.1
mysql-connector-python==8.1.0
joblib==1.3.2
imbalanced-learn==0.11.1
scipy==1.11.1
pytest==7.4.0python-dotenv
